Did some cleanup / refactoring on the httpcache engine , composed some code.
