+++ bundle/src/test/java/com/adobe/acs/commons/httpcache/engine/impl/HttpCacheEngineImplTest.java
@@ -42,7 +42,6 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.Captor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.invocation.InvocationOnMock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.mockito.stubbing.Answer;
